requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
openpyxl
Pillow